finally displaying the changes according to the searchFilter
The most important change is the listToBeFiltered actually should be a reference to the actual ArrayList, changing it is crucial for the update of the view itself. But the list of the songs by genre cannot be a reference variable since it holds all the songs that are continuously being filtered.